+++ bindings/matrix-sdk-ffi/src/timeline/content.rs
@@ -40,7 +40,9 @@ impl From<matrix_sdk_ui::timeline::TimelineItemContent> for TimelineItemContent 
 
             Content::CallInvite => TimelineItemContent::CallInvite,
 
-            Content::RtcNotification => TimelineItemContent::RtcNotification,
+            Content::RtcNotification { call_intent } => TimelineItemContent::RtcNotification {
+                call_intent: call_intent.map(|s| s.to_string()),
+            },
 
             Content::MembershipChange(membership) => {
                 let reason = match membership.content() {
@@ -159,7 +161,9 @@ pub enum TimelineItemContent {
         content: MsgLikeContent,
     },
     CallInvite,
-    RtcNotification,
+    RtcNotification {
+        call_intent: Option<String>,
+    },
     RoomMembership {
         user_id: String,
         user_display_name: Option<String>,