+++ labs/multiverse/src/widgets/room_view/timeline.rs
@@ -139,7 +139,7 @@ fn format_timeline_item(item: &Arc<TimelineItem>, is_thread: bool) -> Option<Lis
                     kind: MsgLikeKind::Poll(_), ..
                 })
                 | TimelineItemContent::CallInvite
-                | TimelineItemContent::RtcNotification => {
+                | TimelineItemContent::RtcNotification { .. } => {
                     return None;
                 }