+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -30,12 +30,14 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Add `call_intent` to `TimelineItemContent::RtcNotification`
+  ([#6412](https://github.com/matrix-org/matrix-rust-sdk/pull/6412))
 - Introduce a `ThreadListService` which offers reactive interfaces for rendering
   and managing the list of threads from a particular room.
-  ([6311](https://github.com/matrix-org/matrix-rust-sdk/pull/6311))
+  ([#6311](https://github.com/matrix-org/matrix-rust-sdk/pull/6311))
 - [**breaking**] Remove the `Room::load_thread_list` in favor of the new `ThreadListService`
-  ([6311](https://github.com/matrix-org/matrix-rust-sdk/pull/6311))
-- Add support for [MSC3489](https://github.com/matrix-org/matrix-spec-proposals/pull/3489)  
+  ([#6311](https://github.com/matrix-org/matrix-rust-sdk/pull/6311))
+- Add support for [MSC3489](https://github.com/matrix-org/matrix-spec-proposals/pull/3489)
   live location sharing through a new `TimelineItemContent::LiveLocation` variant.
 - The internal timeline unique ID may be recycled when an event is deduplicated from the timeline,
   so that embedders can notice that it's the same item and avoid unnecessary re-rendering.